move enums to models folder
