+++ OCKSample/Models/InstallationChannel.swift
@@ -1,0 +1,13 @@
+//
+//  InstallationChannel.swift
+//  OCKSample
+//
+//  Created by Corey Baker on 4/14/23.
+//  Copyright © 2023 Network Reconnaissance Lab. All rights reserved.
+//
+
+import Foundation
+
+enum InstallationChannel: String {
+    case global
+}
