+++ OCKSample/Models/MainViewPath.swift
@@ -1,0 +1,13 @@
+//
+//  MainViewPath.swift
+//  OCKSample
+//
+//  Created by Corey Baker on 4/14/23.
+//  Copyright © 2023 Network Reconnaissance Lab. All rights reserved.
+//
+
+import Foundation
+
+enum MainViewPath {
+    case tabs
+}
