+++ OCKSample/Models/TaskID.swift
@@ -1,0 +1,21 @@
+//
+//  TaskID.swift
+//  OCKSample
+//
+//  Created by Corey Baker on 4/14/23.
+//  Copyright © 2023 Network Reconnaissance Lab. All rights reserved.
+//
+
+import Foundation
+
+enum TaskID {
+    static let doxylamine = "doxylamine"
+    static let nausea = "nausea"
+    static let stretch = "stretch"
+    static let kegels = "kegels"
+    static let steps = "steps"
+
+    static var ordered: [String] {
+        [Self.steps, Self.doxylamine, Self.kegels, Self.stretch, Self.nausea]
+    }
+}
