+++ OCKSample/Models/UserType.swift
@@ -1,0 +1,20 @@
+//
+//  UserType.swift
+//  OCKSample
+//
+//  Created by Corey Baker on 4/14/23.
+//  Copyright © 2023 Network Reconnaissance Lab. All rights reserved.
+//
+
+import Foundation
+
+enum UserType: String, Codable {
+    case patient                           = "Patient"
+    case none                              = "None"
+
+    // Return all types as an array, make sure to maintain order above
+    func allTypesAsArray() -> [String] {
+        return [UserType.patient.rawValue,
+                UserType.none.rawValue]
+    }
+}
